requirements.txt
pybit==5.11.0
python-dotenv==1.1.1